Rename WlsTimeProfile to WlsDistribution; to be refacotred to OnedDistribution
