+++ src/celeritas/Types.cc
@@ -91,17 +91,4 @@ char const* to_cstring(NuclearFormFactorType value)
 }
 
 //---------------------------------------------------------------------------//
-/*!
- * Get a string corresponding to the wavelength shifting time model selection.
- */
-char const* to_cstring(WlsTimeProfile value)
-{
-    static EnumStringMapper<WlsTimeProfile> const to_cstring_impl{
-        "delta",
-        "exponential",
-    };
-    return to_cstring_impl(value);
-}
-
-//---------------------------------------------------------------------------//
 }  // namespace celeritas
