+++ src/celeritas/Types.hh
@@ -207,16 +207,6 @@ enum class NuclearFormFactorType
     gaussian,
     size_
 };
-
-//---------------------------------------------------------------------------//
-//! Optical photon wavelength shifting time model
-enum class WlsTimeProfile
-{
-    delta,  //!< Delta function
-    exponential,  //!< Exponential decay
-    size_
-};
-
 //---------------------------------------------------------------------------//
 //! Cylindrical coordinate indices
 enum class CylAxis
