+++ src/celeritas/io/ImportParameters.hh
@@ -8,9 +8,10 @@
 
 #include <map>
 
-#include "celeritas/Constants.hh"
+#include "corecel/inp/Grid.hh"
 #include "celeritas/Types.hh"
 #include "celeritas/Units.hh"
+#include "celeritas/optical/Types.hh"
 
 #include "ImportUnits.hh"
 
@@ -148,8 +149,10 @@ struct ImportTransParameters
  */
 struct ImportOpticalParameters
 {
-    WlsTimeProfile wls_time_profile{WlsTimeProfile::delta};
-    WlsTimeProfile wls2_time_profile{WlsTimeProfile::delta};
+    using Dist = optical::WlsDistribution;
+
+    Dist wls_time_profile{Dist::delta};
+    Dist wls2_time_profile{Dist::delta};
 };
 
 //---------------------------------------------------------------------------//
