+++ src/celeritas/optical/model/WavelengthShiftModel.hh
@@ -39,7 +39,7 @@ class WavelengthShiftModel : public Model
     {
         ImportModelClass model{ImportModelClass::size_};
         std::vector<ImportWavelengthShift> data;
-        WlsTimeProfile time_profile{WlsTimeProfile::size_};
+        WlsDistribution time_profile{WlsDistribution::size_};
     };
 
   public:
