+++ test/celeritas/optical/WavelengthShift.test.cc
@@ -39,7 +39,7 @@ class WavelengthShiftTest : public InteractorHostBase,
 
     void SetUp() override {}
 
-    void build_model(WlsTimeProfile time_profile)
+    void build_model(WlsDistribution time_profile)
     {
         auto const& data = this->imported_data();
         WavelengthShiftModel::Input input;
@@ -68,7 +68,7 @@ class WavelengthShiftTest : public InteractorHostBase,
 
 TEST_F(WavelengthShiftTest, data)
 {
-    this->build_model(WlsTimeProfile::exponential);
+    this->build_model(WlsDistribution::exponential);
 
     // Test the material properties of WLS
     WlsMaterialRecord wls_record = data_.wls_record[material_id_];
@@ -121,7 +121,7 @@ TEST_F(WavelengthShiftTest, time_profile)
 
     {
         // Test delta time profile
-        this->build_model(WlsTimeProfile::delta);
+        this->build_model(WlsDistribution::delta);
 
         real_type const expected_time
             = dist.time + data_.wls_record[dist.material].time_constant;
@@ -133,7 +133,7 @@ TEST_F(WavelengthShiftTest, time_profile)
     }
     {
         // Test exponential time profile
-        this->build_model(WlsTimeProfile::exponential);
+        this->build_model(WlsDistribution::exponential);
 
         real_type time_ns = dist.time / units::nanosecond;
         Histogram bin(8, {time_ns, time_ns + 4});
@@ -159,7 +159,7 @@ TEST_F(WavelengthShiftTest, time_profile)
 
 TEST_F(WavelengthShiftTest, wls_basic)
 {
-    this->build_model(WlsTimeProfile::exponential);
+    this->build_model(WlsDistribution::exponential);
 
     int const num_samples = 4;
     auto& rng = this->InteractorHostBase::rng();
@@ -199,7 +199,7 @@ TEST_F(WavelengthShiftTest, wls_basic)
 
 TEST_F(WavelengthShiftTest, wls_stress)
 {
-    this->build_model(WlsTimeProfile::exponential);
+    this->build_model(WlsDistribution::exponential);
 
     int const num_samples = 128;
     auto& rng = this->InteractorHostBase::rng();
